CausalTables.jl

The following is the documentation for each function in the CausalTables.jl package.

Base.rand — Method

rand(dgp::DataGeneratingProcess, n::Int)

Generate a random CausalTable using the specified DataGeneratingProcess.

Arguments

dgp::DataGeneratingProcess: The DataGeneratingProcess object defining the causal network and distribution steps.

n::Int: The number of observations to generate.

Returns

ct::CausalTable: The generated CausalTable.

CausalTables.append_dgp_draw! — Method

append_dgp_draw!(name::Symbol, step::UnivariateDistribution, ct::CausalTable, n)

Appends a draw from a distribution to the CausalTable.

Arguments

name::Symbol: The name of the column to be added.

`step: Either the distribution (or a Vector of distributions) from which to draw values, or a function that summarizes neighboring distributions in the causal graph between units.

ct::CausalTable: The CausalTable to which the values will be appended.

n: The number of values to draw.

CausalTables.condensity — Method

condensity(dgp::DataGeneratingProcess, ct::CausalTable, var::Symbol)

Compute the conditional density of a variable in a CausalTable based on a DataGeneratingProcess.

Arguments

dgp::DataGeneratingProcess: The DataGeneratingProcess object representing the data generating process.

ct::CausalTable: The CausalTable object containing the data.

var::Symbol: The name of the variable for which to compute the conditional density.

Returns

The conditional density of the variable in the CausalTable.

CausalTables.get_conditional_distribution — Method

get_conditional_distribution(varfunc::Function, ct)

Compute the conditional distribution of a variable in the DataGeneratingProcess using a given function and a CausalTable.

Arguments

varfunc: The right side of the Pair used to compute the conditional distribution in the DataGeneratingProcess. Either of type Function or NetworkSummary (e.g. NeighborSum).

ct::CausalTable: The CausalTable containing the data.

Returns

The conditional distribution of the variable.

CausalTables.initialize_dgp_step — Method

initialize_dgp_step(step_func::NeighborSum, ct)

Initialize the data generating process (DGP) step by applying the given step_func to the causal table ct.

Arguments

step_func: The function to be applied to the causal table. Either a function or a NetworkSummary (e.g. NeighborSum).

ct: The causal table.

Returns

ct: The modified causal table after applying the step_func.

Distributions.convolve — Method

Distributions.convolve(ds::Vector{T}) where {T <: UnivariateDistribution}

Overload the convolve function to work on a vector of UnivariateDistribution.

Arguments

ds::Vector{T}: A vector of UnivariateDistribution objects.

Returns

output: The result of convolving all the distributions in ds.

CausalTables.CausalTable — Type

CausalTable

A mutable structure that contains a table (tbl), a graph (graph), and a named tuple of summaries (summaries).

CausalTables.CausalTable — Method

CausalTable(tbl)

An alternate constructor for CausalTable that takes a table (tbl) and initializes the graph and summaries.

CausalTables.DataGeneratingProcess — Type

mutable struct DataGeneratingProcess

Mutable struct representing a data generating process.

Fields

networkgen::Function: The function to generate the network.

distgen::Vector{Pair{Symbol, ValidDGPTypes}}: The vector of variable-summary pairs.

CausalTables.DataGeneratingProcess — Method

DataGeneratingProcess(distgen::Vector{Pair{Symbol, ValidDGPTypes}})

Constructs a DataGeneratingProcess object with the given distribution generator.

Arguments

distgen::Vector{Pair{Symbol, ValidDGPTypes}}: A vector of pairs representing the distribution generator.

Returns

DataGeneratingProcess: The constructed DataGeneratingProcess object.

CausalTables.NeighborSum — Type

abstract type NeighborSum <: NetworkSummary

Abstract type representing a summary of neighbor variables.

CausalTables.NeighborSumIn — Type

struct NeighborSumIn <: NeighborSum

Struct representing a summary of incoming neighbor variables.

Fields

var_to_summarize::Symbol: The variable to summarize.

CausalTables.NeighborSumOut — Type

struct NeighborSumOut <: NeighborSum

Struct representing a summary of outgoing neighbor variables.

Fields

var_to_summarize::Symbol: The variable to summarize.

CausalTables.NetworkSummary — Type

abstract type NetworkSummary

Abstract type representing a summary of a network.