Computing Ground Truth Conditional Distributions

Once we've defined a DGP and have some table of data with variables matching those of our DGP, we can compute the ground truth conditional distributions of any variable in a CausalTable (given a corresponding DGP) using the condensity function. This returns a Distribution object from the package Distributions.jl.

Let's see an example. First, we'll define a DGP:

using Graphs
using CausalTables
using Random
using Distributions

dgp = @dgp(
        W ~ Binomial(10, 0.3),
        X ~ (@. Normal(W + 1)),
        A = Graphs.adjacency_matrix(barabasi_albert(length(X), 2)),
        Xs $ Sum(:X, :A),
        Y ~ (@. LogNormal(log(0.2 * Xs + 4), 0.1 * W + 1))
    )

scm = StructuralCausalModel(
    dgp;
    treatment = :X,
    response = :Y,
    confounders = [:W]
)

Now, let's generate some data and compute the ground truth conditional distributions of the variables in the data. Note that if the DGP attempts to summarize a variable with no neighbors in a graph, the resulting conditional distribution will currently be Binomial(0, 0.5), which denotes a point-mass distribution at 0.

Random.seed!(1);
data = rand(scm, 5)
W_distribution = condensity(scm, data, :W)
X_distribution = condensity(scm, data, :X)
Xs_distribution = condensity(scm, data, :Xs)

One can also compute the ground truth conditional mean of a variable in a CausalTable using the conmean function:

Y = conmean(scm, data, :Y)

Ground Truth Conditional Distributions API

CausalTables.condensity — Method

condensity(scm::StructuralCausalModel, ct::CausalTable, name::Symbol)

Compute the conditional density of a variable in a StructuralCausalModel given a CausalTable.

Arguments

scm::StructuralCausalModel: The StructuralCausalModel representing the data generating process.

ct::CausalTable: The CausalTable containing the observed data.

name::Symbol: The variable for which to compute the conditional density.

Returns

The conditional density of the variable var given the observed data.

CausalTables.conmean — Method

conmean(scm::StructuralCausalModel, ct::CausalTable, name::Symbol)

Compute the conditional mean of a variable in a CausalTable based on a DataGeneratingProcess.

Arguments

scm::StructuralCausalModel: The StructuralCausalModel object representing the data generating process.

ct::CausalTable: The CausalTable object representing the data.

name::Symbol: The variable for which to compute the conditional mean.

Returns

An array of conditional means for the specified variable.

CausalTables.convar — Method

convar(scm::StructuralCausalModel, ct::CausalTable, name::Symbol)

Compute the conditional variance of a variable in a CausalTable based on a DataGeneratingProcess.

Arguments

scm::StructuralCausalModel: The StructuralCausalModel object representing the data generating process.

ct::CausalTable: The CausalTable object representing the data.

name::Symbol: The variable for which to compute the conditional mean.

Returns

An array of conditional variances for the specified variable.