Network Summaries

Typically, performing causal inference on a network relies on summarizing the treatment and covariates of each unit's neighbors using some sort of summary function. For example, in a study evaluating the effect of electric vehicle adoption on air pollution, one might model the commuting patterns between counties as a network and evaluate the effect of the sum the number of electric vehicles commuting into each county. The following documents summary measures available in CausalTables.jl and how to summarize a CausalTable

Summarizing a CausalTable

Data wrapped in a CausalTable includes a NamedTuple summaries which describes extra variables represented as summary variables over the network. These summary measures can be computed and added to the table by calling the summarize function on the CausalTable object.

CausalTables.summarize — Function

summarize(x::CausalTable, keep = true)

Summarize a CausalTable by computing the summaries specified in the summaries field of the CausalTable and adding them to the existing data.

Arguments

x::CausalTable: The CausalTable to be summarized.

keep_original::Bool: Whether to return a CausalTable with summarized variables appended (true) or a Tables-compatible NamedTuple with only the summarized variables (false).

Returns

If keep_original is true, a new CausalTable with merged columns, treatment, response, graph, and summaries. Otherwise, NamedTuple of just variable summaries

Example

using Graphs

# Construct a CausalTable
ctbl = CausalTable(
    (A = [1, 2, 3],);
    graph = Graphs.complete_graph(3),
    summaries = (As = Sum(:A), Am = Maximum(:A))
)

# Compute the summaries of the CausalTable
summarize(ctbl)

# output
CausalTable((A = [1, 2, 3], As = [5, 4, 3], Am = [3, 3, 2]), nothing, nothing, nothing, SimpleGraph{Int64}(3, [[2, 3], [1, 3], [1, 2]]), (As = Sum(:A, true), Am 
= Maximum(:A, true)))

Existing Summary Measures

The following lists summary measures currently available off-the-shelf in CausalTables.jl. Examples on their use are provided in Generating Data for Statistical Experiments and Turning data into a CausalTable.

CausalTables.Sum — Type

Sum(var_to_summarize::Symbol; use_inneighbors::Bool = true)

Constructs a Network Summary denoting the sum of var_to_summarize over each unit's neighbors. If var_to_summarize is $X$, then mathematically this computes

\[\sum_{j \in \mathcal{F}_i} X_j\]

where $F_i$ represents the "friends" or neighbors of unit $i$.

Arguments

var_to_summarize::Symbol: The variable to summarize.

use_inneighbors::Bool: Whether to use the in-neighbors or out-neighbors. If true, then the in-neighbors are used. If the graph is undirected, this has no effect.

include_self::Bool: Whether to include the value of var_to_summarize for the unit itself in the sum. Default is true.

Returns

Sum: The constructed Sum object.

CausalTables.Product — Type

Product(var_to_summarize::Symbol; use_inneighbors::Bool = true)

Constructs a Network Summary denoting the product of var_to_summarize over each unit's neighbors. If var_to_summarize is $X$, then mathematically this computes

\[\prod_{j \in \mathcal{F}_i} X_j\]

where $F_i$ represents the "friends" or neighbors of unit $i$.

Arguments

var_to_summarize::Symbol: The variable to summarize.

use_inneighbors::Bool: Whether to use the in-neighbors or out-neighbors. If true, then the in-neighbors are used. If the graph is undirected, this has no effect.

Returns

Product: The constructed Product object.

CausalTables.Maximum — Type

Maximum(var_to_summarize::Symbol; use_inneighbors::Bool = true)

Constructs a Network Summary denoting the maximum of var_to_summarize over each unit's neighbors. If var_to_summarize is $X$, then mathematically this computes

\[\max_{j \in \mathcal{F}_i} X_j\]

where $F_i$ represents the "friends" or neighbors of unit $i$.

Arguments

var_to_summarize::Symbol: The variable to summarize.

use_inneighbors::Bool: Whether to use the in-neighbors or out-neighbors. If true, then the in-neighbors are used. If the graph is undirected, this has no effect.

Returns

Maximum: The constructed Maximum object.

CausalTables.Minimum — Type

Minimum(var_to_summarize::Symbol; use_inneighbors::Bool = true)

Constructs a Network Summary denoting the minimum of var_to_summarize over each unit's neighbors. If var_to_summarize is $X$, then mathematically this computes

\[\min_{j \in \mathcal{F}_i} X_j\]

where $F_i$ represents the "friends" or neighbors of unit $i$.

Arguments

var_to_summarize::Symbol: The variable to summarize.

use_inneighbors::Bool: Whether to use the in-neighbors or out-neighbors. If true, then the in-neighbors are used. If the graph is undirected, this has no effect.

Returns

Minimum: The constructed Minimum object.

CausalTables.Mode — Type

Mode(var_to_summarize::Symbol; use_inneighbors::Bool = true)

Constructs a Network Summary denoting the mode of var_to_summarize over each unit's neighbors – that is, the value occuring most often among units connected to each unit in the network.

Arguments

var_to_summarize::Symbol: The variable to summarize.

use_inneighbors::Bool: Whether to use the in-neighbors or out-neighbors. If true, then the in-neighbors are used. If the graph is undirected, this has no effect.

Returns

Mode: The constructed Mode object.

CausalTables.Friends — Type

Friends(var_to_summarize::Symbol; use_inneighbors::Bool = true)

Constructs a Network Summary denoting the number of neighbors of each unit in the network.

Arguments

use_inneighbors::Bool: Whether to use the in-neighbors or out-neighbors. If true, then the in-neighbors are used. If the graph is undirected, this has no effect.

Returns

Friends: The constructed Friends object.

Defining Your Own Summary Measures

If you want to use a custom summary measure, you can define your own by creating a new type that is a subtype of NetworkSummary. The following example shows how to define a new summary measure that computes the mode of a variable over a unit's neighbors:

mutable struct Mode <: NetworkSummary 
    var_to_summarize::Symbol
    use_inneighbors::Bool
    include_self::Bool
    Mode(var_to_summarize::Symbol; use_inneighbors::Bool = true, include_self = true) = new(var_to_summarize, use_inneighbors, include_self)
end

Then, simply add a new function to the summarize method by defining summarize(x::CausalTable, summary::NetworkSummary) function, but replacing the NetworkSummary with your new summary type. One easy way to define a new summarize function is to use apply_function_over_neighbors. This takes a function taking a vector as input and applies it to the variable specified in second argument, filtered to each unit's neighbors.

In this case, we can use StatsBase.mode to compute the mode of each unit's neighbors.

CausalTables.apply_function_over_neighbors — Function

apply_function_over_neighbors(x::CausalTable, var_to_summarize::Symbol, summary_func::Function; use_inneighbors = true)

Apply a summary function over the neighbors of each row in a CausalTable. Mostly used as a utility to define a summarize function for new NetworkSummary types.

Arguments

x::CausalTable: The CausalTable object.

var_to_summarize::Symbol: The symbol representing the variable in the CausalTable to summarize.

summary_func::Function: The summary function to apply. Should take in a vector as input.

use_inneighbors::Bool: (optional) If true, use the in-neighbors of each row as neighbors. If false, use the out-neighbors. Default is true.

Returns

A Vector containing the summary value for each row.

summarize(x::CausalTable, summary::Mode) = apply_function_over_neighbors(x, summary.var_to_summarize, StatsBase.mode; use_inneighbors = summary.use_inneighbors)