CausalTables.jl

The following is the documentation for each function in the CausalTables.jl package.

Base.rand — Method

rand(dgp::DataGeneratingProcess, n::Int)

Generate a random CausalTable using the specified DataGeneratingProcess.

Arguments

dgp::DataGeneratingProcess: The DataGeneratingProcess object defining the causal network and distribution steps.

n::Int: The number of observations to generate.

Returns

ct::CausalTable: The generated CausalTable.

CausalTables.append_dgp_draw! — Method

append_dgp_draw!(name::Symbol, step::UnivariateDistribution, ct::CausalTable, n)

Appends a draw from a distribution to the CausalTable.

Arguments

name::Symbol: The name of the column to be added.

`step: Either the distribution (or a Vector of distributions) from which to draw values, or a function that summarizes neighboring distributions in the causal graph between units.

ct::CausalTable: The CausalTable to which the values will be appended.

n: The number of values to draw.

CausalTables.condensity — Method

condensity(dgp::DataGeneratingProcess, ct::CausalTable, var::Symbol)

Compute the conditional density of a variable in a CausalTable based on a DataGeneratingProcess.

Arguments

dgp::DataGeneratingProcess: The DataGeneratingProcess object representing the data generating process.

ct::CausalTable: The CausalTable object containing the data.

var::Symbol: The name of the variable for which to compute the conditional density.

Returns

The conditional density of the variable in the CausalTable.

CausalTables.conmean — Method

conmean(dgp::DataGeneratingProcess, ct::CausalTable, var::Symbol)

Compute the conditional mean of a variable in a CausalTable based on a DataGeneratingProcess.

Arguments

dgp::DataGeneratingProcess: The DataGeneratingProcess object representing the data generating process.

ct::CausalTable: The CausalTable object representing the data.

var::Symbol: The variable for which to compute the conditional mean.

Returns

An array of conditional means for the specified variable.

CausalTables.get_conditional_distribution — Method

get_conditional_distribution(varfunc::Function, ct)

Compute the conditional distribution of a variable in the DataGeneratingProcess using a given function and a CausalTable.

Arguments

varfunc: The right side of the Pair used to compute the conditional distribution in the DataGeneratingProcess. Either of type Function or NetworkSummary (e.g. NeighborSum).

ct::CausalTable: The CausalTable containing the data.

Returns

The conditional distribution of the variable.

CausalTables.getcontrols — Method

getcontrols(x::CausalTable)

Selects the control variables from the given CausalTable object x.

Arguments

x::CausalTable: The CausalTable object from which to select the control variables.

keepcausal::Bool: Determines whether to keep the CausalTable wrapping or return a NamedTuple. Default is true.

Returns

A new CausalTable object containing only the control variables.

CausalTables.getgraph — Method

getgraph(x::CausalTable)

Get the graph associated with a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

Returns

The graph associated with the CausalTable.

CausalTables.getresponse — Method

getresponse(x::CausalTable)

Get the response variable from a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

Returns

The response variable from the CausalTable.

CausalTables.getsummaries — Method

getsummaries(x::CausalTable)

Arguments

x::CausalTable: The CausalTable object.

Returns

An array of tables stored in the CausalTable x.

CausalTables.gettable — Method

gettable(x::CausalTable)

Extracts the underlying table from a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

Returns

The underlying table.

CausalTables.gettreatment — Method

gettreatment(x::CausalTable)

Get the treatment column from a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

Returns

The treatment column of the CausalTable.

CausalTables.initialize_dgp_step — Method

initialize_dgp_step(step_func::NeighborSum, ct)

Initialize the data generating process (DGP) step by applying the given step_func to the causal table ct.

Arguments

step_func: The function to be applied to the causal table. Either a function or a NetworkSummary (e.g. NeighborSum).

ct: The causal table.

Returns

ct: The modified causal table after applying the step_func.

CausalTables.summarize — Function

summarize(x::CausalTable, keep = true)

Summarize a CausalTable by merging its columns, treatment, response, graph, and summaries.

Arguments:

x::CausalTable: The CausalTable to be summarized.

keep::Bool: Determines whether to keep the original CausalTable or return a new summarized CausalTable. Default is true.

Returns:

If keep is true, a new CausalTable with merged columns, treatment, response, graph, and summaries.

If keep is false, a NamedTuple with summaries as keys and the corresponding summarized CausalTable as values.

Distributions.convolve — Method

Distributions.convolve(ds::Vector{T}) where {T <: UnivariateDistribution}

Overload the convolve function to work on a vector of UnivariateDistribution.

Arguments

ds::Vector{T}: A vector of UnivariateDistribution objects.

Returns

output: The result of convolving all the distributions in ds.

CausalTables.CausalTable — Type

CausalTable

A mutable structure that contains a table (tbl), a graph (graph), and a named tuple of summaries (summaries).

CausalTables.DataGeneratingProcess — Type

struct DataGeneratingProcess

A mutable struct representing a data generating process.

Fields

networkgen::Function: A function that generates the network structure.

distgen::Vector{Pair{Symbol, ValidDGPTypes}}: A vector of pairs representing the distribution generators for each variable.

treatment::Symbol: The symbol representing the treatment variable.

response::Symbol: The symbol representing the response variable.

controls::Vector{Symbol}: A vector of symbols representing the control variables.

CausalTables.NeighborSum — Type

abstract type NeighborSum <: NetworkSummary

Struct representing a summary of neighbor variables.

CausalTables.NetworkSummary — Type

abstract type NetworkSummary

Abstract type representing a summary of a network.