Base.rand — Method

rand(dgp::DataGeneratingProcess, n::Int)

Generate a random CausalTable using the specified DataGeneratingProcess.

Arguments

dgp::DataGeneratingProcess: The DataGeneratingProcess object defining the causal network and distribution steps.

n::Int: The number of observations to generate.

Returns

ct::CausalTable: The generated CausalTable.

CausalTables.condensity — Method

condensity(dgp::DataGeneratingProcess, ct::CausalTable, var::Symbol)

Compute the conditional density of a variable in a CausalTable based on a DataGeneratingProcess.

Arguments

dgp::DataGeneratingProcess: The DataGeneratingProcess object representing the data generating process.

ct::CausalTable: The CausalTable object containing the data.

var::Symbol: The name of the variable for which to compute the conditional density.

Returns

The conditional density of the variable in the CausalTable.

CausalTables.conmean — Method

conmean(dgp::DataGeneratingProcess, ct::CausalTable, var::Symbol)

Compute the conditional mean of a variable in a CausalTable based on a DataGeneratingProcess.

Arguments

dgp::DataGeneratingProcess: The DataGeneratingProcess object representing the data generating process.

ct::CausalTable: The CausalTable object representing the data.

var::Symbol: The variable for which to compute the conditional mean.

Returns

An array of conditional means for the specified variable.

CausalTables.getcontrols — Method

getcontrols(x::CausalTable)

Selects the control variables from the given CausalTable object x.

Arguments

x::CausalTable: The CausalTable object from which to select the control variables.

keepcausal::Bool: Determines whether to keep the CausalTable wrapping or return a NamedTuple. Default is true.

Returns

A new CausalTable object containing only the control variables.

CausalTables.getgraph — Method

getgraph(x::CausalTable)

Get the graph associated with a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

Returns

The graph associated with the CausalTable.

CausalTables.getresponse — Method

getresponse(x::CausalTable)

Get the response variable from a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

Returns

The response variable from the CausalTable.

CausalTables.getsummaries — Method

getsummaries(x::CausalTable)

Arguments

x::CausalTable: The CausalTable object.

Returns

An array of tables stored in the CausalTable x.

CausalTables.gettable — Method

gettable(x::CausalTable)

Extracts the underlying table from a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

Returns

The underlying table.

CausalTables.gettreatment — Method

gettreatment(x::CausalTable)

Get the treatment column from a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

Returns

The treatment column of the CausalTable.

CausalTables.replace — Method

replace(x::CausalTable; tbl = nothing, treatment = nothing, response = nothing, controls = nothing, graph = nothing, summaries = nothing)

Conviently replace several components of a CausalTable with new values.

Arguments:

x::CausalTable: The CausalTable object to modify.

tbl: The new table to replace the existing table. If nothing, the current table is used.

treatment: The new treatment symbol to replace the existing treatment symbol. If nothing, the current treatment symbol is used.

response: The new response symbol to replace the existing response symbol. If nothing, the current response symbol is used.

controls: The new control symbols to replace the existing control symbols. If nothing, the current control symbols are used.

graph: The new graph to replace the existing graph. If nothing, the current graph is used.

summaries: The new summaries to replace the existing summaries. If nothing, the current summaries are used.

Returns:

CausalTable: The modified CausalTable object.

CausalTables.replacetable — Method

replacetable(x::CausalTable, tbl)

Conveniently replace the underlying table of a CausalTable with a new table.

Arguments

x::CausalTable: The CausalTable object to replace the table for.

tbl: The new table to replace the existing table with.

Returns

A new CausalTable object with the updated table.

CausalTables.setcausalvars! — Method

setcausalvars!(x::CausalTable; treatment=nothing, response=nothing, controls=nothing)

Convenience function for setting new treatment, response, and controls variables for a CausalTable at once.

Arguments:

x::CausalTable: The CausalTable object.

treatment=nothing: The treatment variable.

response=nothing: The response variable.

controls=nothing: The control variables.

CausalTables.setcontrols! — Method

setcontrols!(x::CausalTable, controls::Vector{Symbol})

Set the control variables for a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

controls::Vector{Symbol}: The new control variables to be set.

CausalTables.setresponse! — Method

setresponse!(x::CausalTable, response::Symbol)

Set the response variable for a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

response::Symbol: The symbol representing the new response variable.

CausalTables.settreatment! — Method

settreatment!(x::CausalTable, treatment::Symbol)

Set the treatment variable for a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

treatment::Symbol: The symbol representing the new treatment variable.

CausalTables.summarize — Function

summarize(x::CausalTable, keep = true)

Summarize a CausalTable by merging its columns, treatment, response, graph, and summaries.

Arguments:

x::CausalTable: The CausalTable to be summarized.

keep::Bool: Determines whether to keep the original CausalTable or return a new summarized CausalTable. Default is true.

Returns:

If keep is true, a new CausalTable with merged columns, treatment, response, graph, and summaries.

If keep is false, a NamedTuple with summaries as keys and the corresponding summarized CausalTable as values.

Distributions.convolve — Method

Distributions.convolve(ds::Vector{T}) where {T <: UnivariateDistribution}

Overload the convolve function to work on a vector of UnivariateDistribution.

Arguments

ds::Vector{T}: A vector of UnivariateDistribution objects.

Returns

output: The result of convolving all the distributions in ds.

Tables.subset — Method

subset(x::CausalTable, ind)

Subset a CausalTable x based on the given indices ind. Note that viewhinting is not supported; this function will return a copy of the CausalTable.

Arguments

x: The CausalTable to be subsetted.

ind: The indices to subset the table and graph.

Returns

A new CausalTable with the subsetted table. If the graph attribute is not Nothing, then this function takes the subgraph induced by the subsetted indices using induced

CausalTables.CausalTable — Type

CausalTable

A mutable structure that contains a table (tbl), a graph (graph), and a named tuple of summaries (summaries).

CausalTables.DataGeneratingProcess — Type

struct DataGeneratingProcess

A mutable struct representing a data generating process.

Fields

networkgen::Function: A function that generates the network structure.

distgen::Vector{Pair{Symbol, ValidDGPTypes}}: A vector of pairs representing the distribution generators for each variable.

treatment::Symbol: The symbol representing the treatment variable.

response::Symbol: The symbol representing the response variable.

controls::Vector{Symbol}: A vector of symbols representing the control variables.

CausalTables.NeighborSum — Type

abstract type NeighborSum <: NetworkSummary

Struct representing a summary of neighbor variables.

CausalTables.NetworkSummary — Type

abstract type NetworkSummary

Abstract type representing a summary of a network.