API

The following documents public methods of CausalTables.jl.

Base.rand — Method

rand(scm::StructuralCausalModel, n::Int)

Generate random data from a Structural Causal Model (SCM) using the specified number of samples.

Arguments

scm::StructuralCausalModel: The Structural Causal Model from which to generate data.

n::Int: The number of samples to generate.

Returns

A CausalTable object containing the generated data.

CausalTables.condensity — Method

condensity(scm::StructuralCausalModel, ct::CausalTable, var::Symbol)

Compute the conditional density of a variable in a StructuralCausalModel given a CausalTable.

Arguments

scm::StructuralCausalModel: The StructuralCausalModel representing the data generating process.

ct::CausalTable: The CausalTable containing the observed data.

var::Symbol: The variable for which to compute the conditional density.

Returns

The conditional density of the variable var given the observed data.

CausalTables.conmean — Method

conmean(dgp::DataGeneratingProcess, ct::CausalTable, var::Symbol)

Compute the conditional mean of a variable in a CausalTable based on a DataGeneratingProcess.

Arguments

dgp::DataGeneratingProcess: The DataGeneratingProcess object representing the data generating process.

ct::CausalTable: The CausalTable object representing the data.

var::Symbol: The variable for which to compute the conditional mean.

Returns

An array of conditional means for the specified variable.

CausalTables.getscm — Method

getscm(o::CausalTable)

Get the structural causal model (SCM) of a CausalTable object.

This function merges the column table of the CausalTable object with its arrays.

Arguments

o::CausalTable: The CausalTable object.

Returns

A merged table containing the column table and arrays of the CausalTable object.

CausalTables.replace — Method

replace(o::CausalTable; kwargs...)

Replace the fields of a CausalTable object with the provided keyword arguments.

Arguments

o::CausalTable: The CausalTable object to be replaced.

kwargs...: Keyword arguments specifying the new values for the fields.

Returns

A new CausalTable object with the specified fields replaced.

CausalTables.summarize — Method

summarize(o::CausalTable; keep_original=true)

Summarizes the data in a CausalTable object.

Arguments

o::CausalTable: The CausalTable object to be summarized.

keep_original::Bool: Whether to keep the original data in the resulting CausalTable. Default is true.

Returns

If keep_original is true, a new CausalTable object with the original data merged with the summarized data.

If keep_original is false, a dictionary containing the summarized data.

Distributions.convolve — Method

Distributions.convolve(ds::Vector{T}) where {T <: UnivariateDistribution}

Overload the convolve function to work on a vector of UnivariateDistribution.

Arguments

ds::Vector{T}: A vector of UnivariateDistribution objects.

Returns

output: The result of convolving all the distributions in ds. If ds is empty, will return Binomial(0, 0.5) denoting a point mass at 0.

CausalTables.DataGeneratingProcess — Type

mutable struct DataGeneratingProcess

A struct representing a data generating process.

Fields

names: An array of symbols representing the names of the variables.

types: An array of symbols representing the types of the variables.

funcs: An array of functions representing the generating functions for each variable.

CausalTables.Friends — Type

mutable struct Friends <: NetworkSummary

A NetworkSummary counting the number of connected individuals in an adjacency matrix, also known as the number of "friends"

Fields

matrix::Symbol: The matrix representing the network.

CausalTables.NetworkSummary — Type

abstract type NetworkSummary

Abstract type representing a summary of a network.

CausalTables.StructuralCausalModel — Type

struct StructuralCausalModel

A struct representing a structural causal model (SCM).

Fields

dgp::DataGeneratingProcess: The data generating process associated with the structural causal model.

treatment::Vector{Symbol}: The variables representing the treatment in the structural causal model.

response::Vector{Symbol}: The variables representing the response in the structural causal model.

confounders::Vector{Symbol}: The variables representing the confounders in the structural causal model.

Constructors

StructuralCausalModel(dgp, treatment, response, confounders): Constructs a new StructuralCausalModel object.

Arguments

dgp: The data generating process associated with the structural causal model.

treatment: The variables representing the treatment in the structural causal model.

response: The variables representing the response in the structural causal model.

confounders: The variables representing the confounders in the structural causal model.

CausalTables.Sum — Type

Sum <: NetworkSummary

A NetworkSummary which sums the values of the target variable for each unit connected in the adjacency matrix.

Fields

target::Symbol: The target variable of the network.

matrix::Symbol: The matrix representation of the network.