API

The following documents public methods of CausalTables.jl.

Base.rand — Method

rand(dgp::DataGeneratingProcess, n::Int)

Generate a random CausalTable using the specified DataGeneratingProcess.

Arguments

dgp::DataGeneratingProcess: The DataGeneratingProcess object defining the causal network and distribution steps.

n::Int: The number of observations to generate.

Returns

ct::CausalTable: The generated CausalTable.

CausalTables.apply_function_over_neighbors — Method

apply_function_over_neighbors(x::CausalTable, var_to_summarize::Symbol, summary_func::Function; use_inneighbors = true)

Apply a summary function over the neighbors of each row in a CausalTable. Mostly used as a utility to define a summarize function for new NetworkSummary types.

Arguments

x::CausalTable: The CausalTable object.

var_to_summarize::Symbol: The symbol representing the variable in the CausalTable to summarize.

summary_func::Function: The summary function to apply. Should take in a vector as input.

use_inneighbors::Bool: (optional) If true, use the in-neighbors of each row as neighbors. If false, use the out-neighbors. Default is true.

Returns

A Vector containing the summary value for each row.

CausalTables.condensity — Method

condensity(dgp::DataGeneratingProcess, ct::CausalTable, var::Symbol)

Compute the conditional density of a variable in a CausalTable based on a DataGeneratingProcess.

Arguments

dgp::DataGeneratingProcess: The DataGeneratingProcess object representing the data generating process.

ct::CausalTable: The CausalTable object containing the data.

var::Symbol: The name of the variable for which to compute the conditional density.

Returns

The conditional density of the variable in the CausalTable.

CausalTables.conmean — Method

conmean(dgp::DataGeneratingProcess, ct::CausalTable, var::Symbol)

Compute the conditional mean of a variable in a CausalTable based on a DataGeneratingProcess.

Arguments

dgp::DataGeneratingProcess: The DataGeneratingProcess object representing the data generating process.

ct::CausalTable: The CausalTable object representing the data.

var::Symbol: The variable for which to compute the conditional mean.

Returns

An array of conditional means for the specified variable.

CausalTables.getcontrols — Method

getcontrols(x::CausalTable)

Selects the control variables from the given CausalTable object x.

Arguments

x::CausalTable: The CausalTable object from which to select the control variables.

keepcausal::Bool: Determines whether to keep the CausalTable wrapping or return a NamedTuple. Default is true.

Returns

A new CausalTable object containing only the control variables.

CausalTables.getgraph — Method

getgraph(x::CausalTable)

Get the graph associated with a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

Returns

The graph associated with the CausalTable.

CausalTables.getresponse — Method

getresponse(x::CausalTable)

Get the response variable from a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

Returns

The response variable from the CausalTable.

CausalTables.getsummaries — Method

getsummaries(x::CausalTable)

Arguments

x::CausalTable: The CausalTable object.

Returns

An array of tables stored in the CausalTable x.

CausalTables.gettable — Method

gettable(x::CausalTable)

Extracts the underlying table from a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

Returns

The underlying table.

CausalTables.gettreatment — Method

gettreatment(x::CausalTable)

Get the treatment column from a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

Returns

The treatment column of the CausalTable.

CausalTables.replace — Method

replace(x::CausalTable; tbl = nothing, treatment = nothing, response = nothing, controls = nothing, graph = nothing, summaries = nothing)

Conviently replace several components of a CausalTable with new values.

Arguments:

x::CausalTable: The CausalTable object to modify.

tbl: The new table to replace the existing table. If nothing, the current table is used.

treatment: The new treatment symbol to replace the existing treatment symbol. If nothing, the current treatment symbol is used.

response: The new response symbol to replace the existing response symbol. If nothing, the current response symbol is used.

controls: The new control symbols to replace the existing control symbols. If nothing, the current control symbols are used.

graph: The new graph to replace the existing graph. If nothing, the current graph is used.

summaries: The new summaries to replace the existing summaries. If nothing, the current summaries are used.

Returns:

CausalTable: The modified CausalTable object.

CausalTables.replacetable — Method

replacetable(x::CausalTable, tbl)

Conveniently replace the underlying table of a CausalTable with a new table.

Arguments

x::CausalTable: The CausalTable object to replace the table for.

tbl: The new table to replace the existing table with.

Returns

A new CausalTable object with the updated table.

CausalTables.setcausalvars! — Method

setcausalvars!(x::CausalTable; treatment=nothing, response=nothing, controls=nothing)

Convenience function for setting new treatment, response, and controls variables for a CausalTable at once.

Arguments:

x::CausalTable: The CausalTable object.

treatment=nothing: The treatment variable.

response=nothing: The response variable.

controls=nothing: The control variables.

CausalTables.setcontrols! — Method

setcontrols!(x::CausalTable, controls::Vector{Symbol})

Set the control variables for a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

controls::Vector{Symbol}: The new control variables to be set.

CausalTables.setresponse! — Method

setresponse!(x::CausalTable, response::Symbol)

Set the response variable for a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

response::Symbol: The symbol representing the new response variable.

CausalTables.settreatment! — Method

settreatment!(x::CausalTable, treatment::Symbol)

Set the treatment variable for a CausalTable.

Arguments

x::CausalTable: The CausalTable object.

treatment::Symbol: The symbol representing the new treatment variable.

CausalTables.summarize — Method

summarize(x::CausalTable, keep = true)

Summarize a CausalTable by computing the summaries specified in the summaries field of the CausalTable and adding them to the existing data.

Arguments

x::CausalTable: The CausalTable to be summarized.

keep_original::Bool: Whether to return a CausalTable with summarized variables appended (true) or a Tables-compatible NamedTuple with only the summarized variables (false).

Returns

If keep_original is true, a new CausalTable with merged columns, treatment, response, graph, and summaries. Otherwise, NamedTuple of just variable summaries

Example

using Graphs

# Construct a CausalTable
ctbl = CausalTable(
    (A = [1, 2, 3],);
    graph = Graphs.complete_graph(3),
    summaries = (As = Sum(:A), Am = Maximum(:A))
)

# Compute the summaries of the CausalTable
summarize(ctbl)

# output
CausalTable((A = [1, 2, 3], As = [5, 4, 3], Am = [3, 3, 2]), nothing, nothing, nothing, SimpleGraph{Int64}(3, [[2, 3], [1, 3], [1, 2]]), (As = Sum(:A, true), Am 
= Maximum(:A, true)))

Distributions.convolve — Method

Distributions.convolve(ds::Vector{T}) where {T <: UnivariateDistribution}

Overload the convolve function to work on a vector of UnivariateDistribution.

Arguments

ds::Vector{T}: A vector of UnivariateDistribution objects.

Returns

output: The result of convolving all the distributions in ds.

Tables.subset — Method

subset(x::CausalTable, ind)

Subset a CausalTable x based on the given indices ind. Note that viewhinting is not supported; this function will return a copy of the CausalTable.

Arguments

x: The CausalTable to be subsetted.

ind: The indices to subset the table and graph.

Returns

A new CausalTable with the subsetted table. If the graph attribute is not Nothing, then this function takes the subgraph induced by the subsetted indices using induced

CausalTables.CausalTable — Type

CausalTable

A mutable structure that contains a table (tbl), a graph (graph), and a named tuple of summaries (summaries).

CausalTables.DataGeneratingProcess — Type

struct DataGeneratingProcess

A mutable struct representing a data generating process.

Fields

networkgen::Function: A function that generates the network structure.

distgen::Vector{Pair{Symbol, ValidDGPTypes}}: A vector of pairs representing the distribution generators for each variable.

treatment::Symbol: The symbol representing the treatment variable.

response::Symbol: The symbol representing the response variable.

controls::Vector{Symbol}: A vector of symbols representing the control variables.

CausalTables.Friends — Method

Friends(var_to_summarize::Symbol; use_inneighbors::Bool = true)

Constructs a Network Summary denoting the number of neighbors of each unit in the network.

Arguments

use_inneighbors::Bool: Whether to use the in-neighbors or out-neighbors. If true, then the in-neighbors are used. If the graph is undirected, this has no effect.

Returns

Friends: The constructed Friends object.

CausalTables.Maximum — Method

Maximum(var_to_summarize::Symbol; use_inneighbors::Bool = true)

Constructs a Network Summary denoting the maximum of var_to_summarize over each unit's neighbors. If var_to_summarize is $X$, then mathematically this computes

\[\max_{j \in \mathcal{F}_i} X_j\]

where $F_i$ represents the "friends" or neighbors of unit $i$.

Arguments

var_to_summarize::Symbol: The variable to summarize.

use_inneighbors::Bool: Whether to use the in-neighbors or out-neighbors. If true, then the in-neighbors are used. If the graph is undirected, this has no effect.

Returns

Maximum: The constructed Maximum object.

CausalTables.Minimum — Method

Minimum(var_to_summarize::Symbol; use_inneighbors::Bool = true)

Constructs a Network Summary denoting the minimum of var_to_summarize over each unit's neighbors. If var_to_summarize is $X$, then mathematically this computes

\[\min_{j \in \mathcal{F}_i} X_j\]

where $F_i$ represents the "friends" or neighbors of unit $i$.

Arguments

var_to_summarize::Symbol: The variable to summarize.

use_inneighbors::Bool: Whether to use the in-neighbors or out-neighbors. If true, then the in-neighbors are used. If the graph is undirected, this has no effect.

Returns

Minimum: The constructed Minimum object.

CausalTables.Mode — Method

Mode(var_to_summarize::Symbol; use_inneighbors::Bool = true)

Constructs a Network Summary denoting the mode of var_to_summarize over each unit's neighbors – that is, the value occuring most often among units connected to each unit in the network.

Arguments

var_to_summarize::Symbol: The variable to summarize.

use_inneighbors::Bool: Whether to use the in-neighbors or out-neighbors. If true, then the in-neighbors are used. If the graph is undirected, this has no effect.

Returns

Mode: The constructed Mode object.

CausalTables.NetworkSummary — Type

abstract type NetworkSummary

Abstract type representing a summary of a network.

CausalTables.Product — Method

Product(var_to_summarize::Symbol; use_inneighbors::Bool = true)

Constructs a Network Summary denoting the product of var_to_summarize over each unit's neighbors. If var_to_summarize is $X$, then mathematically this computes

\[\prod_{j \in \mathcal{F}_i} X_j\]

where $F_i$ represents the "friends" or neighbors of unit $i$.

Arguments

var_to_summarize::Symbol: The variable to summarize.

use_inneighbors::Bool: Whether to use the in-neighbors or out-neighbors. If true, then the in-neighbors are used. If the graph is undirected, this has no effect.

Returns

Product: The constructed Product object.

CausalTables.Sum — Method

Sum(var_to_summarize::Symbol; use_inneighbors::Bool = true)

Constructs a Network Summary denoting the sum of var_to_summarize over each unit's neighbors. If var_to_summarize is $X$, then mathematically this computes

\[\sum_{j \in \mathcal{F}_i} X_j\]

where $F_i$ represents the "friends" or neighbors of unit $i$.

Arguments

var_to_summarize::Symbol: The variable to summarize.

use_inneighbors::Bool: Whether to use the in-neighbors or out-neighbors. If true, then the in-neighbors are used. If the graph is undirected, this has no effect.

include_self::Bool: Whether to include the value of var_to_summarize for the unit itself in the sum. Default is true.

Returns

Sum: The constructed Sum object.